app/backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.10.0,<3.0.0
python-multipart>=0.0.18,<1.0.0
httpx>=0.28.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
psutil>=7.0.0,<8.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
